create_release/pyproject.toml
[tool.poetry]
name = "create_release"
version = "0.1.0"
description = "Tool that creates releases and changelog based on commits"
authors = ["Kay Hannay <klinux@hannay.de>"]

[tool.poetry.dependencies]
python = "^3.7"
click = "^7.1.2"
gitpython = "^3.1.11"

[tool.poetry.dev-dependencies]
pytest = "^6.1.2"
pylint = "^2.6.0"
pytest-mock = "^3.3.1"
pytest-cov = "^2.10.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
